Threaded Network Connection
When the user wants to obtain a traffic report, now an AsyncTask is
implemented which handles the network connection to the website for the
HttpPost on a separate thread from the main thread. This should work on
a phone now (needs to be tested). Updating the imageview (with the map
image obtained) needs to be done on the main thread.